chore: enhance installation scripts with Linux checks and package updates 📦
- Add Linux check to install scripts
- Update packages in install scripts
- Add Docker GPG key installation

+++ scripts/install.fish
@@ -18,6 +18,12 @@ if test (id -u) -ne 0
     exit 1
 end
 
+# check if running on Linux
+if test (uname -s | string lower) != linux
+    echo "This script only supports Linux"
+    exit 1
+end
+
 # update the install the nested packages
 apt update && apt install stow -y
